#57 Rename class files to CamelCase
