+++ exasol-driver-lua-0.1.0-1.rockspec
@@ -54,17 +54,17 @@ build = {
     type = "builtin",
     modules = {
         luasqlexasol = "src/luasqlexasol.lua",
-        connection_properties = "src/connection_properties.lua",
-        connection = "src/connection.lua",
+        ConnectionProperties = "src/ConnectionProperties.lua",
+        Connection = "src/Connection.lua",
         constants = "src/constants.lua",
-        cursor_data = "src/cursor_data.lua",
-        cursor = "src/cursor.lua",
-        environment = "src/environment.lua",
-        exasol_websocket = "src/exasol_websocket.lua",
+        CursorData = "src/CursorData.lua",
+        Cursor = "src/Cursor.lua",
+        Environment = "src/Environment.lua",
+        ExasolWebsocket = "src/ExasolWebsocket.lua",
         luws = "src/luws.lua",
         util = "src/util.lua",
-        websocket_datahandler = "src/websocket_datahandler.lua",
-        websocket = "src/websocket.lua",
+        WebsocketDatahandler = "src/WebsocketDatahandler.lua",
+        Websocket = "src/Websocket.lua",
     },
     copy_directories = { "doc" }
 }
