+++ spec/unit/connection_spec.lua
@@ -1,8 +1,8 @@
 ---@diagnostic disable: undefined-global
 -- luacheck: globals describe it before_each after_each
 require("busted.runner")()
-local connection = require("connection")
-local ConnectionProperties = require("connection_properties")
+local connection = require("Connection")
+local ConnectionProperties = require("ConnectionProperties")
 local resultstub = require("resultstub")
 local config = require("config")
 config.configure_logging()
