+++ src/ConnectionProperties.lua
